+++ CHANGELOG.rst
@@ -7,8 +7,18 @@
 Changelog
 =========
 
-0.9.1 (2024-08-xx)
-------------------
+0.10.0 (2024-08-xx)
+-------------------
+
+**New features**
+
+* Add abstract method
+  :meth:`~metalearners.metalearner.MetaLearner.predict_conditional_average_outcomes`
+  to :class:`~metalearners.metalearner.MetaLearner`.
+
+* Implement
+  :meth:`~metalearners.rlearner.RLearner.predict_conditional_average_outcomes`
+  for :class:`~metalearners.rlearner.RLearner`.
 
 **Bug fixes**
 
@@ -20,7 +30,7 @@ Changelog
 
 **New features**
 
-* Add :meth:`metalearners.metalearner.MetaLearner.init_params`.
+* Add :meth:`metalearners.metalearner.MetaLearner.init_args`.
 
 * Add :class:`metalearners.utils.FixedBinaryPropensity`.